backend/requirements.txt
fastapi==0.114.0
numpy==1.24.2
Pillow==10.4.0
pydantic==2.9.0
uvicorn==0.30.5
wandb==0.17.6
keras==3.5.0